+++ src/StateSmith/Input/DrawIo/MxCell.cs
@@ -1,4 +1,6 @@
-﻿namespace StateSmith.Input.DrawIo;
+﻿using System.Collections.Generic;
+
+namespace StateSmith.Input.DrawIo;
 #nullable enable
 
 public class MxCell
@@ -11,9 +13,22 @@ public class MxCell
     public Type type = Type.None;
     public string? source;
     public string? target;
+    public string? style;
+
+    public Dictionary<string, string> styleMap = new();
 
     public MxCell(string id)
     {
         this.id = id;
+    }
+
+    public bool HasMatchingStyle(string key, string value)
+    {
+        if (styleMap.TryGetValue(key, out var readValue))
+        {
+            return value == readValue;
+        }
+
+        return false;
     }
 }